requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4
dj-database-url>=2.2,<3
psycopg[binary]>=3.2,<4
gunicorn>=23,<24
